Refactor tables.
Refactor the SQL tables out of modules and into a more visible central location.

	modified:   scripts/ganga-daemon.py
	modified:   scripts/userdb-update.py
	modified:   src/python/apache_utils.py
	modified:   src/python/services/RequestsDB.py
	new file:   src/python/tables/Requests.py
	new file:   src/python/tables/Users.py
	new file:   src/python/tables/__init__.py
