+++ src/python/tables/__init__.py
@@ -1,0 +1,7 @@
+"""
+Initialise module.
+
+Load all tables at module scope.
+"""
+from .Users import Users
+from .Requests import Requests
